ab_aligned_buffer

Struct OwnedAlignedBuffer

pub struct OwnedAlignedBuffer { /* private fields */ }

Owned aligned buffer for executor purposes.

See SharedAlignedBuffer for a version that can be cheaply cloned, while reusing the original allocation.

Data is aligned to 16 bytes (128 bits), which is the largest alignment required by primitive types and by extension any type that implements TrivialType/IoType.

Implementations

impl OwnedAlignedBuffer

pub fn with_capacity(capacity: u32) -> Self

Create a new instance with at least specified capacity.

NOTE: Actual capacity might be larger due to alignment requirements.

pub fn from_bytes(bytes: &[u8]) -> Self

Create a new instance from provided bytes.

Panics

If bytes.len() doesn’t fit into u32

pub fn as_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [u8]

pub fn as_ptr(&self) -> *const u8

pub fn as_mut_ptr(&mut self) -> *mut u8

pub fn into_shared(self) -> SharedAlignedBuffer

pub fn ensure_capacity(&mut self, capacity: u32)

Ensure capacity of the buffer is at least capacity.

Will re-allocate if necessary.

pub fn copy_from_slice(&mut self, bytes: &[u8])

Will re-allocate if capacity is not enough to store provided bytes.

Panics

If bytes.len() doesn’t fit into u32

pub fn append(&mut self, bytes: &[u8]) -> bool

Will re-allocate if capacity is not enough to store provided bytes.

Returns false if self.len() + bytes.len() doesn’t fit into u32.

pub fn is_empty(&self) -> bool

pub fn len(&self) -> u32

pub fn capacity(&self) -> u32

pub unsafe fn set_len(&mut self, new_len: u32)

Set the length of the useful data to specified value.

Safety

There must be new_len bytes initialized in the buffer.

Panics

If bytes.len() doesn’t fit into u32